% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_fitness.R
\name{evaluate_fitness}
\alias{evaluate_fitness}
\title{Evaluating the fitness  of the population}
\usage{
evaluate_fitness(population)
}
\arguments{
\item{population}{The list of individuals of the population.}
}
\value{
The output expected should be a list of \eqn{f(x)} values
calculated from individuals in the population.
}
\description{
The function described below applies this equation: \deqn{f(x) = x^2 - 4x + 4}
It assesses every individual \eqn{x} of a given population,
to then provide an overview of how that population fits the equation.
}
\examples{
# example of usage
population <- c(1, 3, 0)
# Evaluate fitness
genetic.algo.optimizeR::evaluate_fitness(population)

}
\author{
Dany Mukesha
}
