% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic.algo.optimizeR-package.R
\docType{package}
\name{genetic.algo.optimizeR-package}
\alias{genetic.algo.optimizeR}
\alias{genetic.algo.optimizeR-package}
\title{genetic.algo.optimizeR: What the Package Does (Title Case)}
\description{
R package for learning purposes that implements a genetic algorithm optimization process.
}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})

}
\keyword{internal}
