% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_population.R
\name{initialize_population}
\alias{initialize_population}
\title{Initialize population}
\usage{
initialize_population(population_size, min = -100, max = 100)
}
\arguments{
\item{population_size}{The number of individuals of the population.}

\item{min}{The minimum number an individual could have.}

\item{max}{The maximum number an individual could have.}
}
\value{
The output expected should be a list of individuals of the population
with the size indicated in the input.
}
\description{
The function described below creates a population of individuals(values).
}
\details{
(Note: the values are random and the population should be highly diversified) \cr
The space of x value is kept integer type and on range from 0 to 3, for simplification.
}
\examples{
# example of usage
genetic.algo.optimizeR::initialize_population(population_size = 3, min = 0, max = 3)

}
\author{
Dany Mukesha
}
