% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genetic.algo.optimizeR-package.R
\docType{package}
\name{genetic.algo.optimizeR-package}
\alias{genetic.algo.optimizeR}
\alias{genetic.algo.optimizeR-package}
\title{genetic.algo.optimizeR: Genetic Algorithm Optimization}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

R package for learning purposes that implements a genetic algorithm optimization process.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://danymukesha.github.io/genetic.algo.optimizeR/}
  \item \url{https://github.com/danymukesha/genetic.algo.optimizeR}
  \item Report bugs at \url{https://github.com/danymukesha/genetic.algo.optimizeR/issues}
}

}
\author{
\strong{Maintainer}: Dany Mukesha \email{danymukesha@gmail.com} (\href{https://orcid.org/0009-0001-9514-751X}{ORCID})

}
\keyword{internal}
